requirements.txt
Pillow>=10.0.0
Requests>=2.31.0
